Per issue #10 this commit starts to implement the entire-api.
This move renames the entire.router param on package.json to
entire.backend to help explain that the entire-api is for more
than simply adding routes to the middleware stack.

In this commit I also changed how non-permitted routes are handled.
Before they were responded to with a 401, now they will simply be
skipped. This will allow different permission types to register
different route handelers for the same route.

In cleaning up how routes are handeled I've also manged to limit
the entire-middleware stack to those middleware connected to the
requested permission set. Doing this let me move from .call to
koa-compose and closes #5.